+++ CHANGELOG.md
@@ -13,7 +13,10 @@ and this project adheres to [Semantic Versioning][semver].
 [#274]: https://github.com/openlawlibrary/pygls/issues/274
 
 ### Changed
+- Default behaviour change: uncaught errors are now sent as `showMessage` errors to client.
+  Overrideable in `LanguageServer.report_server_error()`: https://github.com/openlawlibrary/pygls/pull/282
 ### Fixed
+- `_data_recevied()` JSONRPC message parsing errors now caught
 
 ## [0.12.4] - 24/10/2022
 ### Fixed